vision_stack/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drone-vision-stack"
version = "1.0.0"
description = "Fixed-wing drone vision-guidance system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Drone Vision Team"}
]

dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "depthai>=2.20.0",
    "pymavlink>=2.4.0",
    "pyzmq>=22.0.0",
    "pyyaml>=6.0",
    "ultralytics>=8.0.0",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
drone-vision = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP"]

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
